move DB files to allow multiple DBs
