--- persistence/src/commonMain/sqldelight/com/wire/kalium/persistence/kalium_db/ServerConfig.sq
@@ -1,23 +1,0 @@
-CREATE TABLE ServerConfigration (
-    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-    apiBaseUrl TEXT NOT NULL,
-    accountBaseUrl TEXT NOT NULL,
-    webSocketBaseUrl TEXT NOT NULL,
-    blackListUrl TEXT NOT NULL,
-    teamsUrl TEXT NOT NULL,
-    websiteUrl TEXT NOT NULL,
-    title TEXT NOT NULL
-);
-
-deleteById:
-DELETE FROM ServerConfigration WHERE id = ?;
-
-insert:
-INSERT OR IGNORE INTO ServerConfigration(id, apiBaseUrl, accountBaseUrl, webSocketBaseUrl, blackListUrl, teamsUrl, websiteUrl, title)
-VALUES(?, ?, ?, ?, ?, ?, ?, ?);
-
-storedConfig:
-SELECT * FROM ServerConfigration;
-
-getById:
-SELECT * FROM ServerConfigration WHERE id = ?;
