+++ persistence/src/commonMain/user_db/com/wire/kalium/persistence/Assets.sq
