+++ persistence/src/commonMain/user_db/com/wire/kalium/persistence/Metadata.sq
