+++ persistence/src/commonMain/user_db/com/wire/kalium/persistence/Teams.sq
